Added ui for domain expired
