+++ ui/app/rif/pages/domainsDetailPage/domainExpired/domainExpired.css
@@ -1,0 +1,31 @@
+.domain-expired {
+  display: flex;
+  flex-flow: column;
+  padding: 3em;
+  align-items: center;
+}
+
+.domain-expired div {
+  color: #E60000;
+  text-transform: uppercase;
+  font-size: 1.2em;
+  font-weight: bold;
+  margin-top: 2em;
+  text-align: center;
+}
+.domain-expired-icon {
+  font-size: 5em;
+}
+.domain-expired-mark {
+  float: right;
+  margin-right: 1em;
+  width: 5em;
+  background-color: #E60000;
+  color: white;
+  border-radius: 2em;
+  font-size: 0.8em;
+  align-items: center;
+  display: flex;
+  flex-flow: column;
+  padding: 0.3em 4em;
+}
