+++ ui/app/rif/pages/domainsDetailPage/domainsDetailPage.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
 import DomainsDetailActiveScreen from './domainDetailActive/domainDetailActive'
-import DomainExpired from './domainExpired/domainDetailActive'
+import DomainExpired from './domainExpired/domainExpired'
 
 class DomainsDetailScreen extends Component {
 	static propTypes = {
@@ -15,7 +15,10 @@ class DomainsDetailScreen extends Component {
 			{status === 'active' &&
 				<DomainsDetailActiveScreen />
 			}
-			{status !== 'active' &&
+      {status === 'expired' &&
+        <DomainExpired />
+      }
+			{(status !== 'expired' && status !== 'active') &&
 				<div>
 					Domain detail page still in progress for this status!
 				</div>
